+++ src/test/java/blackjack/domain/card_hand/BlackjackCardHandTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card_hand;
 
 import static blackjack.test_util.TestConstants.*;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import blackjack.domain.card.Card;
 
-import blackjack.domain.card_hand.BlackjackCardHand;
 import java.util.List;
 
 import java.util.stream.Stream;
