+++ src/test/java/blackjack/domain/card_hand/CardHandTest.java
@@ -1,9 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.card_hand;
 
 import static blackjack.test_util.TestConstants.*;
 import static org.assertj.core.api.Assertions.assertThat;
-
-import blackjack.domain.card_hand.BlackjackCardHand;
 
 import java.util.List;
