+++ src/test/java/blackjack/domain/deck/DeckTest.java
@@ -1,10 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.deck;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.deck.BlackjackDeck;
 import org.junit.jupiter.api.Test;
 
 class DeckTest {
