+++ src/test/java/blackjack/domain/money/MoneyTest.java
@@ -1,9 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.money;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-
-import blackjack.domain.money.Money;
 
 import java.util.List;
