+++ src/test/java/blackjack/domain/player/PlayerTest.java
@@ -1,6 +1,4 @@
-package blackjack.domain;
-
-import blackjack.domain.player.Player;
+package blackjack.domain.player;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
